moving routes and extensions to seperate files
